Use Utils class instead
