+++ lib/data/process-test-directory/index.js
@@ -7,30 +7,29 @@
 'use strict';
 
 const path = require('path');
-const { Readable } = require('stream');
-
-const csv = require('csv-parser');
 const beautify = require('json-beautify');
 
 const { validate } = require('../../util/error');
-const { reindent } = require('../../util/lines');
 const { Queryable } = require('../../util/queryable');
 const { FileRecordChain } = require('../../util/file-record-chain');
 
 const { parseSupport } = require('../parse-support');
-const { parseTestCSVRowV2 } = require('../parse-test-csv-row');
-const { parseCommandCSVRowV2, flattenObject } = require('../parse-command-csv-row');
+const { parseTestCSVRowV2: parseTestCSVRow } = require('../parse-test-csv-row');
+const {
+  parseCommandCSVRowV2: parseCommandCSVRow,
+  flattenObject,
+} = require('../parse-command-csv-row');
 const {
   createAtCommandTuplesATSettingsTestIdLookupByPresentationNumber,
 } = require('../command-tuples-at-mode-task-lookup');
-
-const { renderHTML: renderCollectedTestHtml } = require('../templates/collected-test.html');
 const { createExampleScriptsTemplate } = require('../example-scripts-template');
 const {
-  logger,
+  Utils,
   isScriptedReferenceRecord,
   createExampleScriptedFile,
   toBuffer,
+  loadScriptsSource,
+  createScriptFiles,
 } = require('./utils');
 
 /**
@@ -45,7 +44,14 @@ const processTestDirectory = async config => {
   const args = config?.args || {};
   const TEST_MODE = !!args.testMode;
 
-  const { log, VALIDATE_CHECK } = logger(!!args.verbose, !!args.validate);
+  // setup from arguments passed to npm script
+  const utils = new Utils({
+    testFormatVersion: 2,
+    isVerbose: !!args.verbose,
+    isValidate: !!args.validate,
+  });
+
+  const { log, VALIDATE_CHECK } = utils.logger;
 
   /** Name of the test plan. */
   const testPlanName = path.basename(directory);
@@ -71,6 +77,8 @@ const processTestDirectory = async config => {
   const testPlanBuildDirectory = path.join(buildTestsDirectory, testPlanName);
   const indexFileBuildOutputPath = path.join(testPlanBuildDirectory, 'index.html');
 
+  utils.testPlanDirectory = testPlanDirectory;
+
   const existingBuildPromise = FileRecordChain.read(buildDirectory, {
     glob: [
       '',
@@ -90,9 +98,10 @@ const processTestDirectory = async config => {
         filepath => FileRecordChain.read(filepath)
       )
     );
-
   const scriptsRecord = testPlanRecord.find('data/js');
   const resourcesRecord = resourcesOriginalRecord.filter({ glob: '{aria-at-*,keys,vrender}.mjs' });
+
+  utils.testPlanRecord = testPlanRecord;
 
   // Filter out reference html files with inline scripts. Files that are not
   // regenerated will be removed from the filesystem.
@@ -344,12 +353,12 @@ const processTestDirectory = async config => {
 
   const [referencesCsv, testsCsv, assertionsCsv, scriptsCsv, ...atCommandsCsvs] = await Promise.all(
     [
-      readCSVFile('data/references.csv', validateReferencesKeys),
-      readCSVFile('data/tests.csv', validateTestsKeys),
-      readCSVFile('data/assertions.csv', validateAssertionsKeys),
-      readCSVFile('data/scripts.csv', validateScriptsKeys),
+      utils.readCSVFile('data/references.csv', validateReferencesKeys),
+      utils.readCSVFile('data/tests.csv', validateTestsKeys),
+      utils.readCSVFile('data/assertions.csv', validateAssertionsKeys),
+      utils.readCSVFile('data/scripts.csv', validateScriptsKeys),
       ...atCommandsCsvFilePaths.map(({ atKey }) =>
-        readCSVFile(`data/${atKey}-commands.csv`, validateCommandsKeys)
+        utils.readCSVFile(`data/${atKey}-commands.csv`, validateCommandsKeys)
       ),
     ]
   );
@@ -360,7 +369,7 @@ const processTestDirectory = async config => {
     commands: atCommandsCsvs[index],
   }));
 
-  const testsParsed = parseTestCSVRowV2({
+  const testsParsed = parseTestCSVRow({
     tests: testsCsv,
     assertions: assertionsCsv,
     scripts: scriptsCsv,
@@ -754,47 +763,6 @@ ${rows}
     });
   }
 
-  // intended to be an internal helper to reduce some code duplication and make logging for csv errors simpler
-  async function readCSVFile(filePath, rowValidator = identity => identity) {
-    let rawCSV = [];
-    try {
-      rawCSV = await readCSV(testPlanRecord.find(filePath));
-    } catch (e) {
-      log.warning(`WARNING: Error reading ${path.join(testPlanDirectory, filePath)}`);
-      return rawCSV;
-    }
-
-    let index = 0;
-    function printError(message) {
-      // line number is index+2
-      log.warning(
-        `WARNING: Error parsing ${path.join(testPlanDirectory, filePath)} line ${
-          index + 2
-        }: ${message}`
-      );
-    }
-    try {
-      const firstRowKeysLength = Object.keys(rawCSV[0]).length;
-      for (; index < rawCSV.length; index++) {
-        const keysLength = Object.keys(rawCSV[index]).length;
-        if (keysLength != firstRowKeysLength) {
-          printError(
-            `column number mismatch, please include empty cells to match headers. Expected ${firstRowKeysLength} columns, found ${keysLength}`
-          );
-        }
-        if (!rowValidator(rawCSV[index])) {
-          printError('validator returned false result');
-          return;
-        }
-      }
-    } catch (err) {
-      printError(err);
-      return;
-    }
-    log(`Successfully parsed ${path.join(testPlanDirectory, filePath)}`);
-    return rawCSV;
-  }
-
   for (const row of referencesCsv) {
     const {
       references: { aria, htmlAam },
@@ -823,7 +791,7 @@ ${rows}
   }
 
   const scriptsSource = loadScriptsSource(scriptsRecord);
-  const commandsParsed = parseCommandCSVRowV2(
+  const commandsParsed = parseCommandCSVRow(
     {
       commands: parsedAtCommandsCsvs,
     },
@@ -832,13 +800,12 @@ ${rows}
   const {
     references: { aria, htmlAam },
   } = supportJson;
-  const referencesParsed = parseReferencesCSV(referencesCsv, { aria, htmlAam });
+  const referencesParsed = utils.parseReferencesCSV(referencesCsv, { aria, htmlAam });
 
   const keyDefs = flattenObject(commandsJson);
-  const keysParsed = parseKeyMap(keyDefs);
+  const keysParsed = utils.parseKeyMap(keyDefs);
   const supportParsed = parseSupport(supportJson);
 
-  // TODO: This causes 'delete' to incorrectly not be recognized as a key, why?
   const keysValidated = validateKeyMap(keysParsed, {
     addKeyMapError(reason) {
       errorCount += 1;
@@ -922,7 +889,7 @@ ${rows}
         reference: referenceQueryable,
         supportAt: supportQueryables.ats,
         example: exampleScriptedFilesQueryable,
-        modeInstructionTemplate: MODE_INSTRUCTION_TEMPLATES_QUERYABLE(supportJson),
+        modeInstructionTemplate: utils.MODE_INSTRUCTION_TEMPLATES_QUERYABLE(supportJson),
       })
     );
   });
@@ -1007,14 +974,14 @@ ${rows}
         .toString()
         .padStart(2, '0');
       const fileName = `test-${presentationNumber}-${collectedTest.info.testId}-${collectedTest.target.at.key}`;
-      return createCollectedTestFile(collectedTest, testPlanBuildDirectory, fileName);
+      return utils.createCollectedTestFile(collectedTest, testPlanBuildDirectory, fileName);
     }),
     ...combinedCollectedTests.map(collectedTest => {
       const presentationNumber = parseInt(collectedTest.info.presentationNumber)
         .toString()
         .padStart(2, '0');
       const fileName = `test-${presentationNumber}-${collectedTest.info.testId}-${collectedTest.target.at.key}`;
-      return createCollectedTestHtmlFile(collectedTest, testPlanBuildDirectory, fileName);
+      return utils.createCollectedTestHtmlFile(collectedTest, testPlanBuildDirectory, fileName);
     }),
   ];
   buildFiles.forEach(file => {
@@ -1091,161 +1058,6 @@ ${rows}
 };
 
 exports.processTestDirectory = processTestDirectory;
-
-/**
- * @param {FileRecord.Record} record
- * @returns {Promise<string[][]>}
- */
-function readCSV(record) {
-  const rows = [];
-  return new Promise(resolve => {
-    Readable.from(record.buffer)
-      .pipe(
-        csv({
-          mapHeaders: ({ header, index }) => {
-            if (header.toLowerCase().includes('\ufeff'))
-              console.error(
-                `Unwanted U+FEFF found for key ${header} at index ${index} while processing CSV.`
-              );
-            return header.replace(/^\uFEFF/g, '');
-          },
-          mapValues: ({ header, value }) => {
-            if (value.toLowerCase().includes('\ufeff'))
-              console.error(
-                `Unwanted U+FEFF found for value in key, value pair (${header}: ${value}) while processing CSV.`
-              );
-            return value.replace(/^\uFEFF/g, '');
-          },
-        })
-      )
-      .on('data', row => {
-        rows.push(row);
-      })
-      .on('end', () => {
-        resolve(rows);
-      });
-  });
-}
-
-/**
- * @param {FileRecordChain} testPlanJS
- * @returns {AriaATParsed.ScriptSource[]}
- */
-function loadScriptsSource(testPlanJS) {
-  return testPlanJS.filter({ glob: ',*.js' }).entries.map(({ name: fileName, text: source }) => {
-    const name = path.basename(fileName, '.js');
-    const modulePath = path.posix.join('scripts', `${name}.module.js`);
-    const jsonpPath = path.posix.join('scripts', `${name}.jsonp.js`);
-
-    /** @type {AriaATFile.ScriptSource} */
-    return {
-      name,
-      source,
-      modulePath,
-      jsonpPath,
-    };
-  });
-}
-
-/**
- * @param {AriaATFile.ScriptSource[]} scripts
- * @param {string} testPlanBuildDirectory
- * @returns {{path: string, content: Uint8Array}[]}
- */
-function createScriptFiles(scripts, testPlanBuildDirectory) {
-  const jsonpFunction = 'scriptsJsonpLoaded';
-
-  return [
-    ...scripts.reduce((files, script) => {
-      const { modulePath, jsonpPath } = script;
-      const oneScript = [script];
-
-      return [...files, ...renderFileVariants(oneScript, modulePath, jsonpPath)];
-    }, []),
-    ...renderFileVariants(scripts, 'scripts.module.js', 'scripts.jsonp.js'),
-  ];
-
-  function renderFileVariants(scripts, modulePath, jsonpPath) {
-    return [
-      {
-        path: path.join(testPlanBuildDirectory, modulePath),
-        content: encodeText(renderModule(scripts).toString()),
-      },
-      {
-        path: path.join(testPlanBuildDirectory, jsonpPath),
-        content: encodeText(renderJsonp(scripts).toString()),
-      },
-    ];
-  }
-
-  function renderJsonp(scripts) {
-    return reindent`window[document.currentScript.getAttribute("jsonpFunction") || "${jsonpFunction}"]({
-  ${scripts.map(renderJsonpExport).join(',\n')}
-});
-`;
-  }
-
-  function renderJsonpExport({ name, source }) {
-    return reindent`${name}(testPageDocument) {
-  ${source.trim()}
-}`;
-  }
-
-  function renderModule(scripts) {
-    return reindent`${scripts.map(renderModuleExport).join('\n\n')}
-`;
-  }
-
-  function renderModuleExport({ name, source }) {
-    return reindent`export function ${name}(testPageDocument) {
-  ${source.trim()}
-}`;
-  }
-}
-
-/**
- * @param {Object<string, string>} keyDefs
- * @returns {AriaATParsed.KeyMap}
- */
-function parseKeyMap(keyDefs) {
-  const keyMap = {};
-  for (const id in keyDefs) {
-    if (id.includes('.')) {
-      const [type, key] = id.split('.');
-      keyMap[key] = { id: key, type, keystroke: keyDefs[id] };
-    }
-  }
-  return keyMap;
-}
-
-/**
- * @param {AriaATCSV.Reference[]} referenceRows
- * @param {AriaATCSV.SupportReference} aria
- * @param {AriaATCSV.SupportReference} htmlAam
- * @returns {AriaATParsed.ReferenceMap}
- */
-function parseReferencesCSV(referenceRows, { aria, htmlAam }) {
-  const refMap = {};
-  for (const { refId: _refId, type: _type, value: _value, linkText: _linkText } of referenceRows) {
-    let refId = _refId?.trim();
-    let type = _type?.trim();
-    let value = _value?.trim();
-    let linkText = _linkText?.trim();
-
-    if (type === 'aria') {
-      value = `${aria.baseUrl}${aria.fragmentIds[value]}`;
-      linkText = `${linkText} ${aria.linkText}`;
-    }
-
-    if (type === 'htmlAam') {
-      value = `${htmlAam.baseUrl}${htmlAam.fragmentIds[value]}`;
-      linkText = `${linkText} ${htmlAam.linkText}`;
-    }
-
-    refMap[refId] = { refId, type, value, linkText };
-  }
-  return refMap;
-}
 
 /**
  * @param {AriaATParsed.Command} commandParsed
@@ -1332,29 +1144,6 @@ function validateKeyMap(keyMap, { addKeyMapError }) {
 
   return keyMap;
 }
-
-const MODE_INSTRUCTION_TEMPLATES_QUERYABLE = support => {
-  const collection = [];
-  support.ats.forEach(at => {
-    const modes = Object.keys(at.settings);
-    modes.forEach(mode => {
-      collection.push({
-        at: at.key,
-        mode: mode,
-        render: [at.defaultConfigurationInstructionsHTML, ...at.settings[mode].instructions],
-      });
-    });
-
-    // Accounting for modeless AT configurations
-    collection.push({
-      at: at.key,
-      mode: 'defaultMode',
-      render: [at.defaultConfigurationInstructionsHTML],
-    });
-  });
-
-  return Queryable.from('modeInstructionTemplate', collection);
-};
 
 /**
  * @param {T} maybeDefined
@@ -1571,48 +1360,5 @@ function collectTestData({ test, command, supportAt, example, modeInstructionTem
     },
     commands: command.commands,
     assertions: test.assertions,
-  };
-}
-
-function encodeText(text) {
-  if (typeof TextEncoder !== 'undefined') {
-    return new TextEncoder().encode(text);
-  }
-  return new Uint8Array(Buffer.from(text).arrayBuffer);
-}
-
-/**
- * @param {AriaATFile.CollectedTest} test
- * @param {string} testPlanBuildDirectory
- * @param {string} fileName
- * @returns {{path: string, content: Uint8Array}}
- */
-function createCollectedTestFile(test, testPlanBuildDirectory, fileName = null) {
-  const testPresentationNumber = parseInt(test.info.presentationNumber).toString().padStart(2, '0');
-  const testFileName =
-    fileName ||
-    `test-${testPresentationNumber}-${test.info.testId}-${test.target.at.key}-${test.target.at.settings}`;
-
-  return {
-    path: path.join(testPlanBuildDirectory, `${testFileName}.collected.json`),
-    content: encodeText(beautify(test, null, 2, 40)),
-  };
-}
-
-/**
- * @param {AriaATFile.CollectedTest} test
- * @param {string} testPlanBuildDirectory
- * @param {string} fileName
- * @returns {{path: string, content: Uint8Array}}
- */
-function createCollectedTestHtmlFile(test, testPlanBuildDirectory, fileName = null) {
-  const testPresentationNumber = parseInt(test.info.presentationNumber).toString().padStart(2, '0');
-  const testFileName =
-    fileName ||
-    `test-${testPresentationNumber}-${test.info.testId}-${test.target.at.key}-${test.target.at.settings}`;
-
-  return {
-    path: path.join(testPlanBuildDirectory, `${testFileName}.collected.html`),
-    content: encodeText(renderCollectedTestHtml(test, `${testFileName}.collected.json`)),
   };
 }
