+++ lib/data/process-test-directory/v1.js
@@ -7,13 +7,9 @@
 'use strict';
 
 const path = require('path');
-const { Readable } = require('stream');
-
-const csv = require('csv-parser');
 const beautify = require('json-beautify');
 
 const { validate } = require('../../util/error');
-const { reindent } = require('../../util/lines');
 const { Queryable } = require('../../util/queryable');
 const { FileRecordChain } = require('../../util/file-record-chain');
 
@@ -21,13 +17,13 @@ const { parseSupport } = require('../parse-support');
 const { parseTestCSVRow } = require('../parse-test-csv-row');
 const { parseCommandCSVRow } = require('../parse-command-csv-row');
 const { createCommandTuplesATModeTaskLookup } = require('../command-tuples-at-mode-task-lookup');
-
-const { renderHTML: renderCollectedTestHtml } = require('../templates/collected-test.html');
 const { createExampleScriptsTemplate } = require('../example-scripts-template');
 const {
-  logger,
-  isScriptedReferenceRecord,
+  Utils,
   createExampleScriptedFile,
+  createScriptFiles,
+  isScriptedReferenceRecord,
+  loadScriptsSource,
   toBuffer,
 } = require('./utils');
 
@@ -43,7 +39,14 @@ const processTestDirectory = async config => {
   const args = config?.args || {};
   const TEST_MODE = !!args.testMode;
 
-  const { log, VALIDATE_CHECK } = logger(!!args.verbose, !!args.validate);
+  // setup from arguments passed to npm script
+  const utils = new Utils({
+    testFormatVersion: 1,
+    isVerbose: !!args.verbose,
+    isValidate: !!args.validate,
+  });
+
+  const { log, VALIDATE_CHECK } = utils.logger;
 
   const validModes = ['reading', 'interaction', 'item'];
 
@@ -71,6 +74,8 @@ const processTestDirectory = async config => {
 
   let backupTestsCsvFile, backupReferencesCsvFile;
 
+  utils.testPlanDirectory = testPlanDirectory;
+
   const existingBuildPromise = FileRecordChain.read(buildDirectory, {
     glob: [
       '',
@@ -88,9 +93,10 @@ const processTestDirectory = async config => {
       FileRecordChain.read(filepath)
     )
   );
-
   const scriptsRecord = testPlanRecord.find('data/js');
   const resourcesRecord = resourcesOriginalRecord.filter({ glob: '{aria-at-*,keys,vrender}.mjs' });
+
+  utils.testPlanRecord = testPlanRecord;
 
   // Filter out reference html files with inline scripts. Files that are not
   // regenerated will be removed from the filesystem.
@@ -455,6 +461,7 @@ ${references}
   /**
    * Create an index file for a local server
    * @param tasks
+   * @param emitFile
    */
   function createIndexFile(tasks, { emitFile }) {
     let rows = '';
@@ -583,40 +590,6 @@ ${rows}
     });
   }
 
-  // intended to be an internal helper to reduce some code duplication and make logging for csv errors simpler
-  async function readCSVFile(filePath, rowValidator = identity => identity) {
-    const rawCSV = await readCSV(testPlanRecord.find(filePath));
-    let index = 0;
-    function printError(message) {
-      // line number is index+2
-      log.warning(
-        `WARNING: Error parsing ${path.join(testPlanDirectory, filePath)} line ${
-          index + 2
-        }: ${message}`
-      );
-    }
-    try {
-      const firstRowKeysLength = Object.keys(rawCSV[0]).length;
-      for (; index < rawCSV.length; index++) {
-        const keysLength = Object.keys(rawCSV[index]).length;
-        if (keysLength != firstRowKeysLength) {
-          printError(
-            `column number mismatch, please include empty cells to match headers. Expected ${firstRowKeysLength} columns, found ${keysLength}`
-          );
-        }
-        if (!rowValidator(rawCSV[index])) {
-          printError('validator returned false result');
-          return;
-        }
-      }
-    } catch (err) {
-      printError(err);
-      return;
-    }
-    log(`Successfully parsed ${path.join(testPlanDirectory, filePath)}`);
-    return rawCSV;
-  }
-
   function validateReferencesKeys(row) {
     if (typeof row.refId !== 'string' || typeof row.value !== 'string') {
       throw new Error('Row missing refId or value');
@@ -677,21 +650,21 @@ ${rows}
   }
 
   const [atCommands, refRows, tests] = await Promise.all([
-    readCSVFile('data/commands.csv', validateCommandsKeys),
-    readCSVFile(backupReferencesCsvFile || 'data/referencesV1.csv', validateReferencesKeys),
-    readCSVFile(backupTestsCsvFile || 'data/testsV1.csv', validateTestsKeys),
+    utils.readCSVFile('data/commands.csv', validateCommandsKeys),
+    utils.readCSVFile(backupReferencesCsvFile || 'data/referencesV1.csv', validateReferencesKeys),
+    utils.readCSVFile(backupTestsCsvFile || 'data/testsV1.csv', validateTestsKeys),
   ]);
 
   for (const row of refRows) {
     refs[row.refId] = row.value.trim();
   }
 
-  const scripts = loadScripts(scriptsRecord);
+  const scripts = loadScriptsSource(scriptsRecord);
 
   const commandsParsed = atCommands.map(parseCommandCSVRow);
   const testsParsed = tests.map(parseTestCSVRow);
-  const referencesParsed = parseReferencesCSV(refRows);
-  const keysParsed = parseKeyMap(keyDefs);
+  const referencesParsed = utils.parseReferencesCSV(refRows);
+  const keysParsed = utils.parseKeyMap(keyDefs);
   const supportParsed = parseSupport(supportJson);
 
   const keysValidated = validateKeyMap(keysParsed, {
@@ -773,7 +746,7 @@ ${rows}
         reference: referenceQueryable,
         example: exampleScriptedFilesQueryable,
         key: keyQueryable,
-        modeInstructionTemplate: MODE_INSTRUCTION_TEMPLATES_QUERYABLE,
+        modeInstructionTemplate: utils.MODE_INSTRUCTION_TEMPLATES_QUERYABLE(),
       })
     );
   });
@@ -785,10 +758,10 @@ ${rows}
       content,
     })),
     ...testsCollected.map(collectedTest =>
-      createCollectedTestFile(collectedTest, testPlanBuildDirectory)
+      utils.createCollectedTestFile(collectedTest, testPlanBuildDirectory)
     ),
     ...testsCollected.map(collectedTest =>
-      createCollectedTestHtmlFile(collectedTest, testPlanBuildDirectory)
+      utils.createCollectedTestHtmlFile(collectedTest, testPlanBuildDirectory)
     ),
   ];
   buildFiles.forEach(file => {
@@ -864,141 +837,6 @@ ${rows}
 exports.processTestDirectory = processTestDirectory;
 
 /**
- * @param {FileRecord.Record} record
- * @returns {Promise<string[][]>}
- */
-function readCSV(record) {
-  const rows = [];
-  return new Promise(resolve => {
-    Readable.from(record.buffer)
-      .pipe(
-        csv({
-          mapHeaders: ({ header, index }) => {
-            if (header.toLowerCase().includes('\ufeff'))
-              console.error(
-                `Unwanted U+FEFF found for key ${header} at index ${index} while processing CSV.`
-              );
-            return header.replace(/^\uFEFF/g, '');
-          },
-          mapValues: ({ header, value }) => {
-            if (value.toLowerCase().includes('\ufeff'))
-              console.error(
-                `Unwanted U+FEFF found for value in key, value pair (${header}: ${value}) while processing CSV.`
-              );
-            return value.replace(/^\uFEFF/g, '');
-          },
-        })
-      )
-      .on('data', row => {
-        rows.push(row);
-      })
-      .on('end', () => {
-        resolve(rows);
-      });
-  });
-}
-
-/**
- * @param {FileRecordChain} testPlanJS
- * @returns {AriaATParsed.ScriptSource[]}
- */
-function loadScripts(testPlanJS) {
-  return testPlanJS.filter({ glob: ',*.js' }).entries.map(({ name: fileName, text: source }) => {
-    const name = path.basename(fileName, '.js');
-    const modulePath = path.posix.join('scripts', `${name}.module.js`);
-    const jsonpPath = path.posix.join('scripts', `${name}.jsonp.js`);
-
-    /** @type {AriaATFile.ScriptSource} */
-    return {
-      name,
-      source,
-      modulePath,
-      jsonpPath,
-    };
-  });
-}
-
-/**
- * @param {AriaATFile.ScriptSource[]} scripts
- * @param {string} testPlanBuildDirectory
- * @returns {{path: string, content: Uint8Array}[]}
- */
-function createScriptFiles(scripts, testPlanBuildDirectory) {
-  const jsonpFunction = 'scriptsJsonpLoaded';
-
-  return [
-    ...scripts.reduce((files, script) => {
-      const { modulePath, jsonpPath } = script;
-      const oneScript = [script];
-
-      return [...files, ...renderFileVariants(oneScript, modulePath, jsonpPath)];
-    }, []),
-    ...renderFileVariants(scripts, 'scripts.module.js', 'scripts.jsonp.js'),
-  ];
-
-  function renderFileVariants(scripts, modulePath, jsonpPath) {
-    return [
-      {
-        path: path.join(testPlanBuildDirectory, modulePath),
-        content: encodeText(renderModule(scripts).toString()),
-      },
-      {
-        path: path.join(testPlanBuildDirectory, jsonpPath),
-        content: encodeText(renderJsonp(scripts).toString()),
-      },
-    ];
-  }
-
-  function renderJsonp(scripts) {
-    return reindent`window[document.currentScript.getAttribute("jsonpFunction") || "${jsonpFunction}"]({
-  ${scripts.map(renderJsonpExport).join(',\n')}
-});
-`;
-  }
-
-  function renderJsonpExport({ name, source }) {
-    return reindent`${name}(testPageDocument) {
-  ${source.trim()}
-}`;
-  }
-
-  function renderModule(scripts) {
-    return reindent`${scripts.map(renderModuleExport).join('\n\n')}
-`;
-  }
-
-  function renderModuleExport({ name, source }) {
-    return reindent`export function ${name}(testPageDocument) {
-  ${source.trim()}
-}`;
-  }
-}
-
-/**
- * @param {Object<string, string>} keyLines
- * @returns {AriaATParsed.KeyMap}
- */
-function parseKeyMap(keyDefs) {
-  const keyMap = {};
-  for (const id in keyDefs) {
-    keyMap[id] = { id, keystroke: keyDefs[id] };
-  }
-  return keyMap;
-}
-
-/**
- * @param {AriaATCSV.Reference[]} referenceRows
- * @returns {AriaATParsed.ReferenceMap}
- */
-function parseReferencesCSV(referenceRows) {
-  const refMap = {};
-  for (const { refId, value } of referenceRows) {
-    refMap[refId] = { refId, value: value.trim() };
-  }
-  return refMap;
-}
-
-/**
  * @param {AriaATParsed.Command} commandParsed
  * @param {object} data
  * @param {Queryable<AriaATParsed.Key>} data.key
@@ -1062,61 +900,6 @@ function validateKeyMap(keyMap, { addKeyMapError }) {
   }
   return keyMap;
 }
-
-const MODE_INSTRUCTION_TEMPLATES_QUERYABLE = Queryable.from('modeInstructionTemplate', [
-  {
-    at: 'jaws',
-    mode: 'reading',
-    render: data => {
-      const altDelete = data.key.where({ id: 'ALT_DELETE' });
-      const esc = data.key.where({ id: 'ESC' });
-      return `Verify the Virtual Cursor is active by pressing ${altDelete.keystroke}. If it is not, exit Forms Mode to activate the Virtual Cursor by pressing ${esc.keystroke}.`;
-    },
-  },
-  {
-    at: 'jaws',
-    mode: 'interaction',
-    render: data => {
-      const altDelete = data.key.where({ id: 'ALT_DELETE' });
-      const insZ = data.key.where({ id: 'INS_Z' });
-      return `Verify the PC Cursor is active by pressing ${altDelete.keystroke}. If it is not, turn off the Virtual Cursor by pressing ${insZ.keystroke}.`;
-    },
-  },
-  {
-    at: 'nvda',
-    mode: 'reading',
-    render: data => {
-      const esc = data.key.where({ id: 'ESC' });
-      return `Ensure NVDA is in browse mode by pressing ${esc.keystroke}. Note: This command has no effect if NVDA is already in browse mode.`;
-    },
-  },
-  {
-    at: 'nvda',
-    mode: 'interaction',
-    render: data => {
-      const insSpace = data.key.where({ id: 'INS_SPACE' });
-      return `If NVDA did not make the focus mode sound when the test page loaded, press ${insSpace.keystroke} to turn focus mode on.`;
-    },
-  },
-  {
-    at: 'voiceover_macos',
-    mode: 'reading',
-    render: data => {
-      const left = data.key.where({ id: 'LEFT' });
-      const right = data.key.where({ id: 'RIGHT' });
-      return `Toggle Quick Nav ON by pressing the ${left.keystroke} and ${right.keystroke} keys at the same time.`;
-    },
-  },
-  {
-    at: 'voiceover_macos',
-    mode: 'interaction',
-    render: data => {
-      const left = data.key.where({ id: 'LEFT' });
-      const right = data.key.where({ id: 'RIGHT' });
-      return `Toggle Quick Nav OFF by pressing the ${left.keystroke} and ${right.keystroke} keys at the same time.`;
-    },
-  },
-]);
 
 /**
  * @param {T} maybeDefined
@@ -1269,51 +1052,5 @@ function collectTestData({ test, command, key, example, modeInstructionTemplate 
     },
     commands: command.commands,
     assertions: test.assertions,
-  };
-}
-
-function encodeText(text) {
-  if (typeof TextEncoder !== 'undefined') {
-    return new TextEncoder().encode(text);
-  }
-  return new Uint8Array(Buffer.from(text).arrayBuffer);
-}
-
-/**
- * @param {AriaATFile.CollectedTest} test
- */
-function createCollectedTestFile(test, testPlanBuildDirectory) {
-  return {
-    path: path.join(
-      testPlanBuildDirectory,
-      `test-${test.info.testId.toString().padStart(2, '0')}-${test.info.task.replace(
-        /\s+/g,
-        '-'
-      )}-${test.target.mode}-${test.target.at.key}.collected.json`
-    ),
-    content: encodeText(beautify(test, null, 2, 40)),
-  };
-}
-
-/**
- * @param {AriaATFile.CollectedTest} test
- * @param {string} testPlanBuildDirectory
- * @returns {{path: string, content: Uint8Array}}
- */
-function createCollectedTestHtmlFile(test, testPlanBuildDirectory) {
-  const testJsonFileName = `test-${test.info.testId
-    .toString()
-    .padStart(2, '0')}-${test.info.task.replace(/\s+/g, '-')}-${test.target.mode}-${
-    test.target.at.key
-  }.collected.json`;
-  return {
-    path: path.join(
-      testPlanBuildDirectory,
-      `test-${test.info.testId.toString().padStart(2, '0')}-${test.info.task.replace(
-        /\s+/g,
-        '-'
-      )}-${test.target.mode}-${test.target.at.key}.collected.html`
-    ),
-    content: encodeText(renderCollectedTestHtml(test, testJsonFileName)),
   };
 }
